requirements.txt
Flask~=2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2==2.9.9
gunicorn==22.0.0
celery==5.4.0
redis==5.0.4
flask-apispec==0.11.4
boto3==1.35.63
pandas==2.0.3
